student-recommendation-system/requirements.txt
# Core dependencies
streamlit>=1.33.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Data processing and analysis
pandas>=2.0.0

# Visualization
plotly>=5.14.0

# Utilities and error handling
tqdm>=4.65.0
anyhow>=0.0.2

